requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
tenacity==9.0.0
tqdm==4.66.5
faiss-cpu==1.9.0.post1
numpy>=2.0.0
ollama==0.3.3
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
pandas==2.2.3
pydantic==2.9.2
# LLM/RAG Dependencies
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
google-genai>=0.8.0
google-ai-generativelanguage>=0.6.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
